requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0

google-generativeai>=0.3.0

groq>=0.4.0

huggingface-hub>=0.20.0

pytest>=7.4.0

ollama>=0.1.0 
